Comments and make unit tests happy
